UNIVERSAL_AUTOPOST/vc_ru_autopost/requirements.txt
playwright>=1.48.0
python-dotenv>=1.0.0
httpx>=0.27.0
